+++ .prettierignore
@@ -4,3 +4,4 @@ tmp/**
 **/types/**
 storybook-static/**
 packages/coverage/**
+packages/design-system-tokens/src/**
