+++ packages/design-system-tokens/src/themes/healthcare-theme.ts
